+++ EL_FitsData/src/lib/FitsDataFixture.cpp
@@ -27,9 +27,9 @@ namespace Test {
 SmallRaster::SmallRaster(long input_width, long input_height) :
     VecRaster<float>({input_width, input_height}),
     width(input_width), height(input_height) {
-  for(int x=0; x<shape[0]; ++x)
-    for(int y=0; y<shape[1]; ++y)
-      operator[]({x, y}) = 0.1 * y + x;
+  for(long x=0; x<shape[0]; ++x)
+    for(long y=0; y<shape[1]; ++y)
+      operator[]({x, y}) = 0.1F * float(y) + float(x);
 }
 
 bool SmallRaster::approx(const Raster<float>& other, float tol) const {
@@ -50,9 +50,9 @@ bool SmallRaster::approx(const Raster<float>& other, float tol) const {
 SmallTable::SmallTable() :
     extname("MESSIER"),
     nums { 45, 7, 31 },
-    radecs { {56.8500, 24.1167}, {268.4667, -34.7928}, {10.6833, 41.2692} },
+    radecs { {56.8500F, 24.1167F}, {268.4667F, -34.7928F}, {10.6833F, 41.2692F} },
     names { "Pleiades", "Ptolemy Cluster", "Andromeda Galaxy" },
-    dists_mags { 0.44, 1.6, 0.8, 3.3, 2900., 3.4 },
+    dists_mags { 0.44, 1.6, 0.8, 3.3, 2900.0, 3.4 },
     num_col ({"ID", "", 1}, nums),
     radec_col ({"RADEC", "deg", 1}, radecs),
     name_col ({"NAME", "", 68}, names), //TODO 68?
